+++ packages/create-dothtml/CHANGELOG.md
@@ -1,5 +1,11 @@
 # create-dothtml
 
+## 6.2.0
+
+### Minor Changes
+
+- Make `h` a formal alias to `html` and fix several issues with SVG rendering.
+
 ## 6.1.0
 
 ### Minor Changes